Initial move of highligher related classes to SharedUx
