+++ src/AccessibilityInsights.SharedUx/Enums/HighlighterType.cs
@@ -1,0 +1,13 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+namespace AccessibilityInsights.SharedUx.Enums
+{
+    /// <summary>
+    /// Types to indicate highlighters
+    /// </summary>
+    public enum HighlighterType
+    {
+        Default,
+        Selected
+    }
+}
