+++ src/AccessibilityInsights.SharedUx/SharedUx.csproj
@@ -281,6 +281,7 @@
       <DependentUpon>PrivacyLearnMore.xaml</DependentUpon>
     </Compile>
     <Compile Include="Enums\FontSize.cs" />
+    <Compile Include="Enums\HighlighterType.cs" />
     <Compile Include="FileBug\BugFileForm.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -310,6 +311,9 @@
     <Compile Include="Enums\ContrastAnalysisType.cs" />
     <Compile Include="Enums\InspectTabType.cs" />
     <Compile Include="Enums\InspectTabMode.cs" />
+    <Compile Include="Highlighting\HighlightAction.cs" />
+    <Compile Include="Highlighting\HighlightImageAction.cs" />
+    <Compile Include="Highlighting\HighlightOverlayAction.cs" />
     <Compile Include="Interfaces\IHierarchyAction.cs" />
     <Compile Include="Controls\ColorChooser.xaml.cs">
       <DependentUpon>ColorChooser.xaml</DependentUpon>
