+++ src/AccessibilityInsights/MainWindowHelpers/EventsMode.cs
@@ -9,6 +9,7 @@ using AccessibilityInsights.Desktop.UIAutomation.EventHandlers;
 using AccessibilityInsights.DesktopUI.Enums;
 using AccessibilityInsights.SharedUx.Dialogs;
 using System.Collections.Generic;
+using AccessibilityInsights.SharedUx.Highlighting;
 
 namespace AccessibilityInsights
 {
