+++ src/AccessibilityInsights/MainWindowHelpers/Initialize.cs
@@ -14,6 +14,7 @@ using System.Windows.Interop;
 using System.Windows.Shell;
 using System.Globalization;
 using AccessibilityInsights.Desktop.Telemetry;
+using AccessibilityInsights.SharedUx.Highlighting;
 
 namespace AccessibilityInsights
 {
