+++ src/AccessibilityInsights/MainWindowHelpers/InspectMode.cs
@@ -4,6 +4,7 @@ using AccessibilityInsights.Enums;
 using AccessibilityInsights.Misc;
 using AccessibilityInsights.Actions;
 using AccessibilityInsights.Actions.Enums;
+using AccessibilityInsights.SharedUx.Highlighting;
 
 namespace AccessibilityInsights
 {
