+++ src/AccessibilityInsights/Modes/SnapshotModeControl.xaml.cs
@@ -27,6 +27,7 @@ using System.Windows.Automation;
 using System.Windows.Controls;
 using System.Windows.Input;
 using System.Windows.Threading;
+using AccessibilityInsights.SharedUx.Highlighting;
 
 namespace AccessibilityInsights.Modes
 {
