+++ src/AccessibilityInsights/Modes/TestModeControl.xaml.cs
@@ -24,6 +24,7 @@ using System.Windows.Input;
 using System.Windows.Threading;
 using static System.FormattableString;
 using System.Diagnostics;
+using AccessibilityInsights.SharedUx.Highlighting;
 
 namespace AccessibilityInsights.Modes
 {
